change Hash (and RawValue) hex-string representation to show the least-significant field element as big-endian hex string (#338)
* change Hash (and RawValue) hex-string representation to show the least-significant field element as big-endian hex string

The motivation is that since some commits ago, the hex representation
was changed from little-endian to big-endian, and when cropping the long
strings of hex (hex representation of byte-arrays), the small values
(224 bits or less) were being represented by `0x00000000...`, which is
indistinguishable from the `0` value.
This commit updates this cropped representation to print the last
characters of the string (the less signifcant bytes of the big-endian
representation), so that for example for the integer `5` the
representation would be `0x...00000005`.